Update tutorials.textile
The statechart diagram for CallHandling example shows event: after 2 s. This event isn't present in any interface... I'm not going to propose a file change as I don't know how to modify the example. Up to now I succesfully writed and executed light switch example...My involvement in this valid technology will be anyhow very shallow as I'm going to program only 2 hours a day and not only with Java: other things to do. I don't know either what is a pull request in GitHub: I should take a tour on it: perhaps.
Пока